Added the initial version of the CLI, sekeletons for profile operation tests, updated configuration options to keep track of the directories for the application, intermediate products, and outputs, fixed an issue in the get_shapefiles helper, created a means of generating random masks for data extraction, expanded on how to install the application, added the ability to set attribute types and edit modes, the setting of output_file paths is more common in order to aid with testing, updated the NCO check to check for each application, Added an object to structure metadata for an NWM output file, Added the objects for processing profile generation, added approved libraries as dependencies, reduced the number of supported python versions, created a new function to subset data since NCO can't, Added skips to unit tests so they will notify that the tests haven't been written but won't constantly fail, and added unit tests

+++ test/test_nco.py
@@ -8,6 +8,7 @@ import tempfile
 import shutil
 import pathlib
 import logging
+import os
 
 import xarray
 
@@ -23,13 +24,22 @@ FORECAST_INTERVAL: int = 1
 logging.basicConfig(level=logging.DEBUG)
 
 
+NPP_TEST_DATA_DIRECTORY: pathlib.Path = pathlib.Path(os.environ['NPP_TEST_DATA_DIRECTORY']) if 'NPP_TEST_DATA_DIRECTORY' in os.environ else None
+
+
 class NCOTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         """
         Set up the test data on disk
         """
-        cls.data_directory = pathlib.Path(tempfile.mkdtemp())
+        if NPP_TEST_DATA_DIRECTORY.is_dir():
+            cls.created_data_directory: bool = False
+            cls.data_directory = NPP_TEST_DATA_DIRECTORY
+        else:
+            cls.created_data_directory: bool = True
+            cls.data_directory = pathlib.Path(tempfile.mkdtemp())
+
         cls.path_to_specification: pathlib.Path = pathlib.Path(__file__).absolute().parent / "specifications" / "short_range.channel_rt.conus.json"
         cls.test_dataset_config: Dataset = deserialize(Dataset, cls.path_to_specification)
         cls.input_files: typing.Sequence[pathlib.Path] = cls.test_dataset_config.generate_netcdf(
@@ -44,38 +54,47 @@ class NCOTest(unittest.TestCase):
         """
         Remove test data from disk
         """
-        shutil.rmtree(cls.data_directory)
+        if cls.created_data_directory:
+            shutil.rmtree(cls.data_directory)
 
     def test_merge_files(self):
         target_file: pathlib.Path = self.data_directory / "merged_data.nc"
-        nco.merge_files_by_variable(files=self.input_files, output_file=target_file)
+        nco.merge_files(files=self.input_files, output_file=target_file)
 
         merged_dataset: xarray.Dataset = xarray.open_dataset(target_file)
         input_dataset: xarray.Dataset = xarray.open_dataset(self.input_files[0])
 
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_keep_only_variables(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_remove_variables(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_transform_variable(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_apply_mask_by_file(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_apply_or_modify_attribute(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_rename_variable(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_rename_dimension(self):
         self.assertEqual(True, False)  # add assertion here
 
+    @unittest.skip("This has not been implemented yet")
     def test_reorder_dimensions(self):
         self.assertEqual(True, False)  # add assertion here
